+++ R/tt_pos_and_access.R
@@ -295,23 +295,43 @@ setMethod(
     if (obj_name(tt) == path[1]) {
       path <- path[-1]
     }
-    cur <- tt
-    curpath <- path
-    while (length(curpath > 0)) {
-      kids <- tree_children(cur)
-      curname <- curpath[1]
-      if (curname == "@content") {
-        cur <- content_table(cur)
-      } else if (curname %in% names(kids)) {
-        cur <- kids[[curname]]
-      } else {
-        stop("Path appears invalid for this tree at step ", curname)
-      }
-      curpath <- curpath[-1]
-    }
-    cur
+
+    # Extract sub-tables from the tree
+    .extract_through_path(tt, path)
   }
 )
+
+# Recursive helper function to retrieve sub-tables from the tree
+.extract_through_path <- function(cur_tbl, cur_path, no_stop = FALSE) {
+  while (length(cur_path > 0)) {
+    kids <- tree_children(cur_tbl)
+    curname <- cur_path[1]
+    if (curname == "@content") {
+      cur_tbl <- content_table(cur_tbl)
+    } else if (curname %in% names(kids)) {
+      cur_tbl <- kids[names(kids) == curname]
+
+      # Case where there are more than one tree sub node with identical names
+      if (length(cur_tbl) > 1 && length(cur_path) > 1) {
+        cur_tbl <- sapply(cur_tbl, function(cti) .extract_through_path(cti, cur_path[-1], no_stop = TRUE))
+        found_values <- !sapply(cur_tbl, is.null)
+        cur_tbl <- cur_tbl[found_values]
+        if (sum(found_values) == 1) {
+          cur_tbl <- cur_tbl[[1]]
+        }
+        cur_path <- cur_path[1]
+      } else {
+        cur_tbl <- cur_tbl[[1]] # Usual case (only one matching value)
+      }
+    } else if (!no_stop) {
+      stop("Path appears invalid for this tree at step ", curname)
+    } else {
+      return(NULL)
+    }
+    cur_path <- cur_path[-1]
+  }
+  cur_tbl
+}
 
 #' @note Setting `NULL` at a defined path removes the corresponding sub-table.
 #'