+++ tests/testthat/test-subset-access.R
@@ -588,3 +588,40 @@ test_that("tt_at_path and cell_values work with values even if they differ in na
   names(rdf$path[[2]]) <- c("a", "b")
   expect_silent(tt_at_path(tbl, rdf$path[[2]]))
 })
+
+test_that("tt_at_path works with identical split names", {
+  # Regression test #864
+  adsl <- ex_adsl
+  adsl$flag <- sample(c("Y", "N"), nrow(adsl), replace = TRUE)
+
+  afun <- function(x, ...) rcell(label = "Flagged Pop. Count", sum(x == "Y"))
+
+  lyt <- basic_table() %>%
+    analyze("flag", afun = afun) %>%
+    split_rows_by("flag", split_fun = keep_split_levels("Y"), child_labels = "hidden") %>%
+    split_rows_by("SEX") %>%
+    analyze("BMRKR1")
+
+  tbl <- build_table(lyt, adsl)
+
+  expect_equal(
+    tt_at_path(tbl, c("root", "flag", "Y")),
+    tree_children(tree_children(tbl)[[2]])[[1]]
+  )
+
+  # Even with deeper branching
+  lyt <- basic_table() %>%
+    split_rows_by("flag", split_fun = keep_split_levels("Y"), child_labels = "hidden") %>%
+    split_rows_by("SEX", split_fun = keep_split_levels("U")) %>%
+    analyze("BMRKR1") %>%
+    split_rows_by("flag", split_fun = keep_split_levels("Y"), child_labels = "hidden") %>%
+    split_rows_by("SEX", split_fun = keep_split_levels("U")) %>%
+    analyze("AGE")
+
+  tbl <- build_table(lyt, adsl)
+
+  expect_equal(
+    names(tt_at_path(tbl, c("root", "flag", "Y", "SEX", "U"))),
+    rep("flag", 2)
+  )
+})